VAULT-21400 test coverage
